fixed warnings

+++ src/media_tools/mpd.c
@@ -1969,7 +1969,7 @@ retry_import:
 					elt = gf_list_get(pe->element.playlist.elements, k);
 					GF_SAFEALLOC(se, GF_MPD_SegmentTimelineEntry);
 					dur = elt->duration_info * rep->segment_list->timescale;
-					se->duration = dur;
+					se->duration = (u32) dur;
 					se->start_time = start_time;
 					start_time += dur;
 					gf_list_add(rep->segment_list->segment_timeline->entries, se);
@@ -4777,10 +4777,10 @@ GF_Err gf_mpd_split_adaptation_sets(GF_MPD *mpd)
 
 				//serialize
 				gf_mpd_print_adaptation_set(set, f, GF_FALSE, 0);
-				size = gf_ftell(f);
+				size = (u32) gf_ftell(f);
 				data = gf_malloc(size+1);
 				gf_fseek(f, 0, SEEK_SET);
-				size = fread(data, 1, size, f);
+				size = (u32) fread(data, 1, size, f);
 				data[size]=0;
 				blob.data = data;
 				blob.size = size;
